split _vtk, integration of mesh, improve dumphdf
