+++ cfdtools/vtk/__init__.py
@@ -1,1 +1,2 @@
-from cfdtools.vtk._vtk import vtkMesh, vtkList
+from cfdtools.vtk._vtk import vtkMesh
+from cfdtools.vtk._vtkList import vtkList
