+++ tests/test_0_h5.py
@@ -8,13 +8,14 @@ def test_WriteOpen(builddir):
     for i, dtype in enumerate(_hdf5._available_types):
         pname = builddir / f"test{i}.h5"
         f = hdf5.h5File(pname)
-        f.open(mode="w", datatype=dtype)
+        f.open(mode="w", datatype=dtype, version=0) # version is dummy but needed
         f.close()
         g = hdf5.h5File(pname)
         g.open() # default is 'r'
         a = g._h5file.attrs
         assert a['cfdtools_version'] == __version__
         assert a['cfd_datatype'] == dtype
+        assert a['data_version'] == 0
         assert g.datatype == dtype
         g.printinfo()
         g.close()
